backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
pymongo==4.3.3
numpy==1.24.2
pandas==1.5.3
pillow==9.4.0
torch==2.0.0
torchvision==0.15.1
python-dotenv==1.0.0
opencv-python==4.7.0.72
matplotlib==3.7.1
seaborn==0.12.2
google-generativeai==0.1.0
reportlab==3.6.12
textblob==0.17.1
werkzeug==2.2.3
firebase-admin==6.1.0
feedparser==6.0.10 